Move the mapping logic to its own type
